refactor: updated slot imports across the codebase (according to the new slots' file structure)
