+++ chatsky/slots/__init__.py
@@ -1,1 +1,2 @@
-from chatsky.slots.slots import GroupSlot, ValueSlot, RegexpSlot, RegexpGroupSlot, FunctionSlot
+from chatsky.slots.slots import RegexpSlot, RegexpGroupSlot, FunctionSlot
+from chatsky.slots.base_slots import GroupSlot, ValueSlot
